refactor: InboundAdapter 경로 변경 및 AppConfig 객체 지역변수 전환
