+++ src/main/java/racingcar/adapter/inbound/InboundAdapter.java
@@ -1,4 +1,4 @@
-package racingcar.application.port.inbound;
+package racingcar.adapter.inbound;
 
 public interface InboundAdapter {
     void run();
